implements parseFloat instead of parseInt for lat lng

+++ src/reports/dtos/get-estimate.dto.ts
@@ -27,11 +27,11 @@ export class GetEstimateDto {
   @Max(1000000)
   mileage: number;
 
-  @Transform(({ value }) => parseInt(value))
+  @Transform(({ value }) => parseFloat(value))
   @IsLongitude()
   lng: number;
 
-  @Transform(({ value }) => parseInt(value))
+  @Transform(({ value }) => parseFloat(value))
   @IsLatitude()
   lat: number;
 }
